♻️ 7BWJQY task: split scenario command helpers
